requirements.txt
streamlit == 1.52.2
pandas
intersystems-irispython
sqlalchemy == 2.0.43
sqlalchemy-iris == 0.16.0
